Support unwelding divergent attributes at vertices.
Fixes OpenCTM OBJ loader to work with non-indexed OBJ files by using a map of unique vertex signatures. This fixes the "cracked" appearance of models (due to texture seams) when using this code.

The issue is that OBJs might have faces of the form:

1/1/1 2/2/2 3/3/3
1/2/1 2/4/2 3/5/7

where vertices can share normal and texture data in ways that are not compatible with OpenGL, which requires each vertex to have a single index that uniquely specifies each attribute.  The solution is to have a unique map of all vertex signatures and choose indicies for vertices from that.

Change-Id: I589690a73705dfb565fe0f8d1103fd7793f2f065